refactor: split monolithic chat_client into modular components
- Extracted ModelManager for model fetching, parsing, and selection
- Extracted ApiClient for API communication and retry logic
- Extracted ToolExecutor for tool call execution (standard and fallback)
- Extracted CommandHandler for slash command processing
- Updated CMakeLists.txt to include new source files
- Updated README.md with refactoring details and expanded project documentation

The refactoring improves maintainability by separating concerns into focused modules, reducing the chat_client.cpp file from 1333 lines to 67 lines while preserving all functionality.
